Multiple fixes
- `write_scen_data` made silent modifications to the file column of the
scenarios and stored them in the wrong directory (legacy leftover)
- `write_scen_data` is now deprecated and its funtionality is now in
`storeScenarios`
- `storeScenarios` now makes required modifications to the scenarios
data frame, most importantly adding this fucking 'fut' thing. This
means, that we need to be aware of this when we finally remove this
crap, together with 'full' and other bullshit. But for the time being,
'full' is required and it needs to have an entry with 'fut'...
- tests have been adjusted accordingly