Refactor of caching functionality
- The old caching system was too naive
- It would either read or write the cache and thus, information was lost
  from time to time

The new functions allow us to synchronize the cache by using timestamps
beside simply restoring it from disk to memory or writing it from memory
to disk.

Timestamps are saved per value.

Furthermore, `listCache` generates a nice output for the user to get an
overview of what is actually stored in the cache.